+++ x/evm/types/precompiles.go
@@ -16,6 +16,7 @@ const (
 	BankPrecompileAddress         = "0x0000000000000000000000000000000000000804"
 	GovPrecompileAddress          = "0x0000000000000000000000000000000000000805"
 	SlashingPrecompileAddress     = "0x0000000000000000000000000000000000000806"
+	EvidencePrecompileAddress     = "0x0000000000000000000000000000000000000807"
 )
 
 // AvailableStaticPrecompiles defines the full list of all available EVM extension addresses.
@@ -32,4 +33,5 @@ var AvailableStaticPrecompiles = []string{
 	BankPrecompileAddress,
 	GovPrecompileAddress,
 	SlashingPrecompileAddress,
+	EvidencePrecompileAddress,
 }